3 Libraries and Frameworks/8 GraphQL/2_requirements.txt
fastapi
strawberry-graphql
uvicorn
pytest